%% main
\name{OmicCircos-package}
\alias{OmicCircos}
\title{
OmicCircos package
}
\description{
 OmicCircos is for generating high-quality circular maps for omic data.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}
\references{
%%  ~~ possibly secondary sources and usages ~~
 OmicCircus: an R package for simple and circular visualization of omic data (in press)
}
%%\keyword{key}
