%%
\name{TCGA.BC.cnv.2k.60}
\alias{TCGA.BC.cnv.2k.60}
\docType{data}
\title{
 sub copy number data of TCGA Breast Cancer 
}
\description{
 This data set is the sub DNA copy number data of TCGA Breast Cancer.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



