%%
\name{TCGA.BC.fus.lab}
\alias{TCGA.BC.fus.lab}
\docType{data}
\title{
 gene fusion (partly) of TCGA Breast Cancer 
}
\description{
 This data set is the fusion gene symbols of TCGA Breast Cancer.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



