%%
\name{TCGA.BC.fus}
\alias{TCGA.BC.fus}
\docType{data}
\title{
 gene fusion (partly) of TCGA Breast Cancer 
}
\description{
 This data set is the gene fusion data of TCGA Breast Cancer.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



