%%
\name{TCGA.BC.gene.exp.2k.60}
\alias{TCGA.BC.gene.exp.2k.60}
\docType{data}
\title{
 sub gene expression data of TCGA Breast Cancer 
}
\description{
 This data set is the sub gene expression data of TCGA Breast Cancer.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



