%%
\name{TCGA.BC.sample60}
\alias{TCGA.BC.sample60}
\docType{data}
\title{
 sub types of TCGA Breast Cancer sub data set
}
\description{
 This data set is the sub types of TCGA Breast Cancer sub data set 
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



