%%
\name{TCGA.BC_Her2_cnv_exp}
\alias{TCGA.BC_Her2_cnv_exp}
\docType{data}
\title{
 the pvalues of the association between the copy number and gene expression of sub TCGA Breast Cancer data set 
}
\description{
 This data set is the pvalues of the association between the copy number and gene expression of sub TCGA Breast Cancer data set.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



