%%
\name{TCGA.PAM50_genefu_hg18}
\alias{TCGA.PAM50_genefu_hg18}
\docType{data}
\title{
 PAM50 gene list (hg18) 
}
\description{
 This data set is the PAM 50 gene list (hg18).
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



