%%
\name{UCSC.chr.colors}
\alias{UCSC.chr.colors}
\docType{data}
\title{
 chromosome banding colors from UCSC Genome Browser 
}
\description{
 chromosome banding colors from UCSC Genome Browser 
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



