%%
\name{UCSC.hg18.chr}
\alias{UCSC.hg18.chr}
\docType{data}
\title{
 human hg18 cytogenetics table from UCSC Genome Browser
}
\description{
 This data set is the human hg18 cytogenetics data by UCSC.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



