%%
\name{UCSC.hg18}
\alias{UCSC.hg18}
\docType{data}
\title{
 human hg18 mapping reference by UCSC cytogenetics table
}
\description{
 This data set is the human hg18 mapping reference by UCSC cytogenetics table.
 It was genereted by hg18.chr data set using seg.angle.po function.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



