%%
\name{UCSC.hg19.chr}
\alias{UCSC.hg19.chr}
\docType{data}
\title{
 human hg19 cytogenetics table from UCSC Genome Browser
}
\description{
 This data set is the human hg19 cytogenetics data by UCSC.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



