%%
\name{UCSC.hg19.mac3.chr}
\alias{UCSC.hg19.mac3.chr}
\docType{data}
\title{
 human hg19 and monkey rhemac3  cytogenetics tables from UCSC Genome Browser
}
\description{
 This data set is human hg19 and monkey rhemac3 cytogenetics tables from UCSC Genome Browser.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



