%%
\name{UCSC.hg19.mac3}
\alias{UCSC.hg19.mac3}
\docType{data}
\title{
 seg.angle.po output of human hg19 and monkey rhemac3 by UCSC Genome Browser
}
\description{
 This data set is the seg.angle.po output of human hg19 and monkey rhemac3 by UCSC Genome Browser.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



