%%
\name{UCSC.hg19.mm10.chr}
\alias{UCSC.hg19.mm10.chr}
\docType{data}
\title{
 human hg19 and mouse mm10 cytogenetics tables from UCSC Genome Browser
}
\description{
 This data set is human hg19 and mouse mm10 cytogenetics tables from UCSC Genome Browser.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



