%%
\name{UCSC.hg19.mm10}
\alias{UCSC.hg19.mm10}
\docType{data}
\title{
 seg.angle.po output of human hg19 and mouse mm10 by UCSC Genome Browser
}
\description{
 This data set is the seg.angle.po output of the human hg19 and mouse mm10 by UCSC Genome Browser.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



