%%
\name{UCSC.hg19}
\alias{UCSC.hg19}
\docType{data}
\title{
 human hg19 mapping reference by UCSC cytogenetics table 
}
\description{
 This data set is the human hg19 mapping reference by UCSC cytogenetics table.
 It was genereted by hg19.chr data set using seg.angle.po function.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



