%%
\name{UCSC.hs_cyto_mm}
\alias{UCSC.hs_cyto_mm}
\docType{data}
\title{
 genome comparison of human hg19 and mouse mm10 by UCSC Genome Browser
}
\description{
 This data set is the genome comparison of the human hg19 and mouse mm10 by UCSC Genome Browser.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



