%%
\name{UCSC.mm10.chr}
\alias{UCSC.mm10.chr}
\docType{data}
\title{
 mouse mm10 cytogenetics table from UCSC Genome Browser
}
\description{
 This data set is the mouse mm10 cytogenetics data by UCSC.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



