%%
\name{UCSC.mm10}
\alias{UCSC.mm10}
\docType{data}
\title{
 mouse mm10 mapping reference by UCSC cytogenetics table 
}
\description{
 This data set is the mouse mm10 mapping reference by UCSC cytogenetics table.
 It was genereted by mm10.chr data set using seg.angle.po function.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



