%%
\name{UCSC.mm9.chr}
\alias{UCSC.mm9.chr}
\docType{data}
\title{
 mouse mm9 cytogenetics table from UCSC Genome Browser 
}
\description{
 This data set is the mouse mm9 cytogenetics data by UCSC.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



