%%
\name{UCSC.mm9}
\alias{UCSC.mm9}
\docType{data}
\title{
 mouse mm9 mapping reference by UCSC cytogenetics table 
}
\description{
 This data set is the mouse mm9 mapping reference by UCSC cytogenetics table.
 It was genereted by mm9.chr data set using seg.angle.po function.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



