%%
\name{UCSC_cytoband_hg19_Mac3}
\alias{UCSC_cytoband_hg19_Mac3}
\docType{data}
\title{
 combined cytogenetics data in human hg18 and Macaca mulatta Mac3
}
\description{
 This data set is the combined cytogenetics data in human hg18 and 
 Macaca mulatta Mac3
}
\author{
%%  ~~ possibly secondary sources and usages ~~
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}



