%%
\name{UCSC_cytoband_hg19_mm10}
\alias{UCSC_cytoband_hg19_mm10}
\docType{data}
\title{
 combined cytogenetics data in human hg18 and mouse mm10
}
\description{
 This data set is the combined cytogenetics data in human hg18 and 
 mouse mm10
}
\author{
%%  ~~ possibly secondary sources and usages ~~
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}



